Add interpolated geojson categories for each hour
Everything adjusted to new data & index structure.
Analysis takes about 1-2 minutes to complete.

During data import 4 station that contained outlying average values
are removed now as it didn't make sense for them to have higher values.
Probably some high error values that were included in the calculation of the average.

The geojsons should be kept in the repository for the webapp to access,
as the connection with the geoserver did not work properly yet.